requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
coverage==6.4.4
Deprecated==1.2.13
Django==4.0.6
django-debug-toolbar==3.5.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
sqlparse==0.4.2
uritemplate==4.1.1
wrapt==1.14.1
